+++ src/assets/index.ts
@@ -1,20 +1,18 @@
-export { default as defaultImageBlack } from './user-profile-default-black-100x100.png';
-export { default as defaultImageGray } from './user-profile-default-gray-100x100.png';
-export { default as signOutDropDown } from './btn-logout-dropdown-w-contents.png';
-export { default as defaultMainPlan } from './main-background-default-plan-125x101.png';
-export { default as logoColor } from './img-traduler-logo-color-4x.png';
-export { default as logoWhite } from './img-traduler-logo-white-4x.png';
-export { default as NotFoundImg } from './img-not-found.png';
 export { default as LoadingGif } from './loader-all-color.gif';
 export { default as profileDefaultGray } from './profile-default-gray.svg';
 export { default as profileDefaultBlack } from './profile-default-black.svg';
 export { default as tradulerLogo } from './traduler-logo_white-blue.svg';
-export { default as errorImg } from './ImgError.png';
-export { default as landingMockup } from './img-landing-mockup1-3x.png';
-export { default as landingMain } from './img-landing-main-fullscreen.png';
-export { default as landingGroupOne } from './landing-group-photo-01-4x.png';
-export { default as landingGroupTwo } from './landing-group-photo-02-4x.png';
-export { default as landingGroupThree } from './landing-group-photo-03-4x.png';
-export { default as landingMockupTwo } from './img-landing-mockup2-2.5x.png';
-export { default as signUpBG } from './purple-sky-photo.png';
-export { default as signInBG } from './photo-sky-4x.png';
+
+export { default as signInBG } from './img-login-left-photo-sky-4x.webp';
+export { default as signUpBG } from './img-register-left-photo.webp';
+export { default as landingMain } from './img-landing-main-fullscreen.webp';
+export { default as landingMockup } from './img-landing-mockup1-3x.webp';
+export { default as landingGroupOne } from './landing-group-photo-01-4x.webp';
+export { default as landingGroupTwo } from './landing-group-photo-02-4x.webp';
+export { default as landingGroupThree } from './landing-group-photo-03-4x.webp';
+export { default as landingMockupTwo } from './img-landing-mockup2-2.5x.webp';
+export { default as logoColor } from './img-traduler-logo-color-4x.webp';
+export { default as logoWhite } from './img-traduler-logo-white-4x.webp';
+export { default as errorImg } from './ImgError.webp';
+export { default as defaultMainPlan } from './main-background-default-plan-125x101.webp';
+export { default as defaultImageGray } from './user-profile-default-gray-100x100.webp';
