+++ src/types/import-webp.d.ts
@@ -1,0 +1,4 @@
+declare module '*.webp' {
+  const content: string;
+  export default content;
+}
